commit_message

+++ reward-facade/src/main/java/com/ruoyi/reward/facade/dto/UserDTO.java
@@ -5,22 +5,20 @@ import lombok.Data;
 import java.io.Serializable;
 import java.util.Date;
 
+/**
+ * @author sunflower
+ */
 @Data
 public class UserDTO implements Serializable {
     /**
-    *
-    *
-    * t_user_detail.id
-    
+     * t_user_detail.id
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
-    private Integer id;
+    private Long id;
 
     /**
-    *
-    *
-    * t_user_detail.user_id
+     * t_user_detail.user_id
      * 用户唯一标识
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -28,9 +26,7 @@ public class UserDTO implements Serializable {
     private String userId;
 
     /**
-    *
-    *
-    * t_user_detail.nickname
+     * t_user_detail.nickname
      * 昵称
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -38,9 +34,7 @@ public class UserDTO implements Serializable {
     private String nickname;
 
     /**
-    *
-    *
-    * t_user_detail.name
+     * t_user_detail.name
      * 用户姓名
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -48,9 +42,7 @@ public class UserDTO implements Serializable {
     private String name;
 
     /**
-    *
-    *
-    * t_user_detail.gender
+     * t_user_detail.gender
      * 用户性别
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -58,9 +50,7 @@ public class UserDTO implements Serializable {
     private String gender;
 
     /**
-    *
-    *
-    * t_user_detail.id_card
+     * t_user_detail.id_card
      * 身份证号
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -68,9 +58,7 @@ public class UserDTO implements Serializable {
     private String idCard;
 
     /**
-    *
-    *
-    * t_user_detail.mobile_no
+     * t_user_detail.mobile_no
      * 手机号
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -78,9 +66,7 @@ public class UserDTO implements Serializable {
     private String mobileNo;
 
     /**
-    *
-    *
-    * t_user_detail.birthday
+     * t_user_detail.birthday
      * 生日yyyyMMdd
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -88,9 +74,7 @@ public class UserDTO implements Serializable {
     private String birthday;
 
     /**
-    *
-    *
-    * t_user_detail.gmt_create
+     * t_user_detail.gmt_create
      * 创建时间
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -98,19 +82,14 @@ public class UserDTO implements Serializable {
     private Date gmtCreate;
 
     /**
-    *
-    *
-    * t_user_detail.create_by
-    
+     * t_user_detail.create_by
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private String createBy;
 
     /**
-    *
-    *
-    * t_user_detail.gmt_modified
+     * t_user_detail.gmt_modified
      * 修改时间
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -118,19 +97,14 @@ public class UserDTO implements Serializable {
     private Date gmtModified;
 
     /**
-    *
-    *
-    * t_user_detail.modified_by
-    
+     * t_user_detail.modified_by
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private String modifiedBy;
 
     /**
-    *
-    *
-    * t_user_detail.user_source
+     * t_user_detail.user_source
      * 用户注册来源
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -138,9 +112,7 @@ public class UserDTO implements Serializable {
     private String userSource;
 
     /**
-    *
-    *
-    * t_user_detail.user_referee
+     * t_user_detail.user_referee
      * 用户推荐人
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -148,29 +120,21 @@ public class UserDTO implements Serializable {
     private String userReferee;
 
     /**
-    *
-    *
-    * t_user_detail.user_referee_ids
-    
+     * t_user_detail.user_referee_ids
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private String userRefereeIds;
 
     /**
-    *
-    *
-    * t_user_detail.marker
-    
+     * t_user_detail.marker
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private String marker;
 
     /**
-    *
-    *
-    * t_user_detail.reg_date
+     * t_user_detail.reg_date
      * 注册时间
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -178,19 +142,14 @@ public class UserDTO implements Serializable {
     private Date regDate;
 
     /**
-    *
-    *
-    * t_user_detail.level
-    
+     * t_user_detail.level
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private Integer level;
 
     /**
-    *
-    *
-    * t_user_detail.avatar_url
+     * t_user_detail.avatar_url
      * 用户头像地址
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -198,9 +157,7 @@ public class UserDTO implements Serializable {
     private String avatarUrl;
 
     /**
-    *
-    *
-    * t_user_detail.reg_url
+     * t_user_detail.reg_url
      * 注册页面路径
      *
      * @mbg.generated 2020-05-04 16:25:45
@@ -208,20 +165,14 @@ public class UserDTO implements Serializable {
     private String regUrl;
 
     /**
-    *
-    *
-    * t_user_detail.password
-    
+     * t_user_detail.password
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
     private String password;
 
     /**
-    *
-    *
-    * t_user_detail.user_name
-    
+     * t_user_detail.user_name
      *
      * @mbg.generated 2020-05-04 16:25:45
      */
